feat(shared): add shared package with types and utilities
- Introduced a new shared package for common types and utilities used across frontend and backend.
- Added ServerConfig interface for backend configuration.
- Defined ClientToServerEvents and ServerToClientEvents interfaces for Socket.IO communication.
- Created LinemodeState interface to represent TELNET LINEMODE capabilities.
- Implemented TypeScript configuration for the shared package.
- Updated root package.json to include the shared workspace and adjusted build scripts.
