--- backend/src/core/middleware/types/server-config.ts
@@ -1,3 +1,0 @@
-export interface ServerConfig {
-  socketNamespace: string;
-}
