--- frontend/src/app/core/mud/index.ts
@@ -1,2 +1,0 @@
-export { MudClientComponent } from './components/mud-client/mud-client.component';
-export { MudService } from './services/mud.service';
