--- frontend/src/app/features/serverconfig/types/server-config.ts
@@ -1,3 +1,0 @@
-export interface ServerConfiguration {
-  socketNamespace: string;
-}
