--- frontend/src/app/features/sockets/index.ts
@@ -1,2 +1,0 @@
-export { SocketsService } from './sockets.service';
-export type { LinemodeState } from './types/linemode-state';
