--- frontend/src/app/features/sockets/types/client-to-server-events.ts
@@ -1,6 +1,0 @@
-export interface ClientToServerEvents {
-  mudConnect: (initialViewPort: { columns: number; rows: number }) => void;
-  mudDisconnect: () => void;
-  mudInput: (data: string) => void;
-  mudViewportSize: (columns: number, rows: number) => void;
-}
