--- frontend/src/app/features/sockets/types/linemode-state.ts
@@ -1,9 +1,0 @@
-export interface LinemodeState {
-  mode: number;
-  edit: boolean;
-  trapsig: boolean;
-  softTab: boolean;
-  literalEcho: boolean;
-  forwardMask: number[];
-  forwardMaskDescription: string;
-}
