--- frontend/src/app/shared/index.ts
@@ -1,2 +1,0 @@
-export { WithRequired } from './types/with-required';
-export { SecureString, isSecureString } from './types/secure-string';
