+++ frontend/src/app/shared/types/secure-string.ts
@@ -1,11 +1,3 @@
 export type SecureString = {
   value: string;
 };
-
-export function isSecureString(
-  obj: string | SecureString,
-): obj is SecureString {
-  return (
-    !(typeof obj === 'string') && (obj as SecureString).value !== undefined
-  );
-}
