+++ shared/src/index.ts
@@ -1,0 +1,4 @@
+export { ServerConfig } from './config/server-config.js';
+export { ClientToServerEvents } from './sockets/client-to-server-events.js';
+export { LinemodeState } from './sockets/linemode-state.js';
+export { ServerToClientEvents } from './sockets/server-to-client-events.js';
